requirements.txt
requests==2.32.3
pyodbc==5.2.0
lxml==5.3.0
python-dotenv==1.0.1